Crate mithril_doc

Commands to generate a markdown documentation for the command line.

Re-exports

pub use mithril_doc_derive;

Structs

FieldDoc

Information to document a field

GenerateDocCommands

Generate documentation

StructDoc

Information about the struct.

Traits

Documenter

Extractor for struct without Default trait.

DocumenterDefault

Extractor for struct with Default trait.

Derive Macros

Documenter

To extract doc from a struct.

DocumenterDefault

To extract doc from a struct with Default trait.